remove lsp booster and misc tidying
the idea of the booster is very cool, though unfortunately the impact does not seem worth the trade of various things breaking at times (this time, lsp-mode with the suggested wrap).

elsewhere, ls is now gls, so the overwrite is no longer needed.

pet no longer starts eglot on py files.

and, a nix lsp.